자기 글만 지우기

+++ src/main/java/com/caprj1/controller/BoardController.java
@@ -54,9 +54,11 @@ public class BoardController {
 
     // 삭제
     @PostMapping("delete")
-    public String delete(Integer id, RedirectAttributes rttr) {
-        service.remove(id);
+    public String delete(Integer id, Authentication authentication) {
 
+        if (service.hasAccess(id, authentication)) {
+            service.remove(id);
+        }
         return "redirect:/";
     }
